Started working on the raft protocol so we can have a distributed appended only log.
The goal isn't to have the fastest implementation but to have one that
is easy to use and simple.  Should still be faster than most
implementations.  The goal is to have each node able to handle 50,000
write message a second.

Started working on a universal messaging library so we can have a
architecture similar to finance.